add controls to main workbench input
